[Wpf] move label text markup handling to Util
* Convert LabelBackend.SetFormattedText() to an SWC.TextBlock
  extension method and move it to Util.cs
* Add IWpfLinkNavigator interface for backends handling link clicks inside
  a formatted text block
